feature(js): covert widgets JS library to an AMD module
Widgets JS library is no longer unnecessarily loaded on each page. The library
has been converted to a reusable elgg/widgets AMD module
